Cant Start new game mode
Im trying to start working on the New Game mode Josh but i cant understand how all ur systems work. could u just make it so when u click on the tactics mode that it does the name screen then puts the player into the tactics scene so i can work on it. Plz and thanks ya mad genius